requirements.txt
asgiref==3.2.7
asn1crypto==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
clarifai-grpc==6.6.0
click==7.1.2
cloudinary==1.20.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
flake8-polyfill==1.0.2
flake8-quotes==3.0.0
Flask==1.1.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.19.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
google-cloud-vision==0.38.1
googleapis-common-protos==1.51.0
grpcio==1.29.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
olefile==0.46
opencv-python-headless==4.4.0.42
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pep8-naming==0.10.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.12.2
psycopg2==2.8.5
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pytesseract==0.3.3
pytest==5.4.1
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==18.1.1
requests==2.23.0
rsa==4.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
zipp==2.2.0
